Move the definition of the top-namespace methods to the root file
That's where anyone would expect to find their definition.
