+++ lib/solidus_social/engine.rb
@@ -8,12 +8,13 @@ require 'omniauth-amazon'
 require 'deface'
 require 'coffee_script'
 require 'spree/core'
+require 'solidus_social/facebook_omniauth_strategy_ext'
 
 module SolidusSocial
   class Engine < Rails::Engine
     include SolidusSupport::EngineExtensions::Decorators
 
-    isolate_namespace Spree
+    isolate_namespace ::Spree
 
     engine_name 'solidus_social'
 
@@ -38,22 +39,6 @@ module SolidusSocial
         # unloaded so that it is available to devise when loading routes
         load USER_DECORATOR_PATH
       end
-    end
-  end
-
-  def self.configured_providers
-    ::Spree::SocialConfig.providers.keys.map(&:to_s)
-  end
-
-  def self.init_providers
-    ::Spree::SocialConfig.providers.each do |provider, credentials|
-      setup_key_for(provider, credentials[:api_key], credentials[:api_secret])
-    end
-  end
-
-  def self.setup_key_for(provider, key, secret)
-    Devise.setup do |config|
-      config.omniauth provider, key, secret, setup: true
     end
   end
 end
